+++ apps/trackers/jira/query.py
@@ -36,6 +36,7 @@ from .constants import (
     JIRA_EMBARGO_SECURITY_LEVEL_NAME,
     JIRA_INTERNAL_SECURITY_LEVEL_NAME,
     PS_ADDITIONAL_FIELD_TO_JIRA,
+    TrackersAppSettings,
 )
 
 logger = logging.getLogger(__name__)
@@ -176,13 +177,14 @@ class OldTrackerJiraQueryBuilder(TrackerQueryBuilder):
     to generate general tracker save query
     """
 
-    def __init__(self, instance):
+    def __init__(self, instance, settings: Optional[TrackersAppSettings] = None):
         """
         init stuff
         """
         self.instance = instance
         self._query = None
         self._comment = None
+        self.settings = settings or TrackersAppSettings()
 
     @cached_property
     def impact(self):
@@ -874,7 +876,14 @@ class TrackerJiraQueryBuilder(OldTrackerJiraQueryBuilder):
     def generate_downstream_component(self):
         field_name = "Downstream Component Name"
         _, field_id = self.field_check_and_get_values_and_id(field_name)
-        self._query["fields"][field_id] = self.most_important_affect.ps_component
+        component = self.most_important_affect.ps_component
+        if (
+            self.settings.prefer_purls
+            and self.ps_module.ps_product.is_middleware
+            and (purl := self.most_important_affect.purl)
+        ):
+            component = purl
+        self._query["fields"][field_id] = component
 
     def generate_upstream_component(self):
         # TODO: Every time the components change in the flaw, the trackers must be updated as well.